+++ resources/views/user/payment-request.blade.php
@@ -16,10 +16,29 @@
 
                     <div class="mb-3">
                         <label for="">Method</label>
-                        <select class="form-select" name="method">
-                            <option value="Digital Wallet">Digital Wallet (Rp 6.000 minimum revenue)</option>
-                            <option value="Bank">Bank (Rp 100.000 minimum revenue)</option>
-                        </select>
+                        <input class="form-control" name="method" list="method-list" value="{{ old('method') }}">
+                        <datalist id="method-list">
+                            <option>Digital Wallet - OVO</option>
+                            <option>Digital Wallet - Gopay</option>
+                            <option>Digital Wallet - ShopeePay</option>
+                            <option>Digital Wallet - DANA</option>
+                            <option>Pulsa - Telkomsel</option>
+                            <option>Pulsa - XL</option>
+                            <option>Pulsa - Axis</option>
+                            <option>Pulsa - Three</option>
+                            <option>Pulsa - Indosat</option>
+                            <option>Pulsa - SmartFren</option>
+                            <option>Bank - BCA</option>
+                            <option>Bank - BRI</option>
+                            <option>Bank - BNI</option>
+                            <option>Bank - Mandiri</option>
+                            <option>Bank - BTPN</option>
+                        </datalist>
+                        <span class="form-text">
+                            <div>Digital Wallet (Rp 6.000 minimum revenue)</div>
+                            <div>Pulsa (Rp 6.000 minimum revenue)</div>
+                            <div>Bank (Rp 100.000 minimum revenue)</div>
+                        </span>
                         @error('method')
                         <span class="form-text text-danger">
                             {{ $errors->first('method')  }}
@@ -29,14 +48,26 @@
 
                     <div class="mb-3">
                         <label for="">Destination</label>
-                        <input class="form-control" type="text" name="destination" placeholder="081123456999_GOPAY" value="{{ old('destination') }}">
+                        <input class="form-control" type="text" name="destination" placeholder="081123456999" value="{{ old('destination') }}">
                         <span class="form-text">
-                            <div>Digital Wallet Format: 081123456999_GOPAY/DANA/OVO</div>
-                            <div>Bank Format: 0881233_BRI/BNI/BCA:AN_Jojo Rizo</div>
+                            Insert Number of you wallet
                         </span>
                         @error('destination')
                         <span class="form-text text-danger">
                             {{ $errors->first('destination')  }}
+                        </span>
+                        @enderror
+                    </div>
+
+                    <div class="mb-3">
+                        <label for="">Identity</label>
+                        <input class="form-control" type="text" name="identity" placeholder="AN: Ahmad" value="{{ old('identity') }}">
+                        <span class="form-text">
+                            <div>Insert wallet identity</div>
+                        </span>
+                        @error('identity')
+                        <span class="form-text text-danger">
+                            {{ $errors->first('identity')  }}
                         </span>
                         @enderror
                     </div>